Src: Rename TreeItemTypes.h to TreeItemTypes.hpp
